Modified Attribute::delete to work with classic and 64-offset files
NetCDF-4 files automatically switch between define and data mode as needed,
but the classic and 64-bit offset modes do not. This means that if you
open a classic file in append mode (which seems to start in data mode)
and try to delete an attribute, you get the `NC_ENOTINDEFINE` error.

The simple way to address this is to try to delete the attribute, check
if you get that "not in define mode" error, and if so, switch into define
mode long enough to delete the attribute, then switch back to data mode.
This works, but has the downside that if you need to delete multiple
attributes, it will switch back into data mode unnecessarily in between.
Since that does some work to finalize the header, that's somewhat inefficient.

The better solution would be to have all operations check if they get the
"not in define mode" or "not in data mode" errors and switch into the mode
they need, but to not switch back when done. That way switching is handled
automatically, but only when needed. If the user is careful to do all of their
define-mode operations at once, they can save on the overhead of switching
back and forth.